Hibernate Brought in (File -> Project Structure -> Facets -> Add Hibernate. JPA comes in as SpringDataJPA. Once Hibernate is in, the is a 'Persistence' icon in the bottom left of Intellij. The Right-Click the item there, in this case 'shopTrawlerrest' and at the end of the list is 'Generate from Database Schema' select the database. (you will be pointing the wizard java>org.fyp>model so that all the generated java files will be saved in there.
Here a couple of things had to happen...

// Prior to starting this process I added 'Model', 'Controller' and 'Repository' package folders to org.fyp

The naming conventions of the tables had to change. In the MySQL database they're all plural, but that's not ideal for Java objects (i.e. Visits changes to Visit etc), this can be done at this point.

Once you try to move forward following that, there are a number of errors. These are each case that a 'SET' was defined in the database, it will be set as an object. Get around this by changing 'object' to 'String'. There is one exception and that is for the YOB (Year of Birth variable)... in this case I went with Integer.

All everything in the 'model' folder now auto-generated.