JPA requirement (empty constructor) added to each entity. Added constructor to each entity to facilitate simple creation of new entities from data loaded from csv files. The methods are just place holders for now to eliminate simple syntax errors.
JPA annotations empty. ... even though Spring-data-jpa is part of the pom, it did not bring in jpa. JPA added as a facet to the project. and Hibernate is the provider. The entities were deleted and rebuilt again, this time with JPA annotations included.

model/BaseEntity Class added to hold common methods used by all entities.
Each entity refined as 'extends BaseEntity'