Field Added to Visit (userCredited)
This field was added because of the functionality of awarding customers points based on dwell/visit times.

In the event that the customer falls just short of the time amount to avail of a visit (another issue could be losing connectivity for a few minutes). Then they should be able to earn the points associated with the easier. So, until a user has been accreditted with points for a visit, every visit will accumulated until that point... and once they are awarded points, the Visit(s) that provided it now have a 'userCredited' tag notifying that the user has been awarded points for the visit.