Saving a registration caused some problems.
email addresses cause problems for pathvariables in a rest ewndpoint.

unless the emailaddress is placed at the end of a list of path variables, it will appear as 'NaN'. Additionally, if the regex characters ':.+' are not included, part of the email address is chopped after the first period in the address (@gmail.com becomes @gmail). See example of a Rest endpoint which now works.

@RequestMapping(value = "/register/Retailer/{firstname}/{surname}/{password}/{userType}/{phone}/{retailerid}/{gender}/{yob}/{emailAddress:.+}",