CORS problems highlighted a number of other issues.
Spring Security modifies to
a) remove 'REALM' entry.
b)             http.csrf().disable()
                   .cors().disable()
                   .authorizeRequests()
                   and
c)         /* To allow Pre-flight [OPTIONS] request from browser */
           @Override
           public void configure(WebSecurity web) throws Exception {
               web.ignoring().antMatchers(HttpMethod.OPTIONS, "/**");
           }

 used to solve the CORS issues on the server end.

Using ShoppingCentre as a test and just changing one field

After a http.put attempt, json threw up several conversion errors related to the back references. Since the app was retrieving the data as required to satisfy the relations, these back refernces were changed to jsonIgnore.

Also, a put attempt needs to be able to set the id to null, this is not posisble on a primitive type (int), this was change to Integer. The other entities will be changhed accordingly.

    public void update(@RequestBody ShoppingCentre shoppingCentre, @PathParam("id") Integer id  )

 Cors on the client end;
 This will be covered in a similar commit on the angular app.

 Note: THis just uses inMemory login, this will be changed to use the UserRepository.